% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_seq.R
\docType{data}
\name{chip_seq}
\alias{chip_seq}
\title{chip_seq}
\format{An object of class \code{matrix} with 1000 rows and 2 columns.}
\usage{
chip_seq
}
\description{
A matrix of two columns. Each column is a replicate of ChIP-seq data. The observations are the read counts from genomic locations.
}
\keyword{datasets}
