% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.binormal.R
\name{d.binormal}
\alias{d.binormal}
\title{density of binomial distribution with equal mean and sigma on both dimensions}
\usage{
d.binormal(z.1, z.2, mu, sigma, rho)
}
\arguments{
\item{z.1}{pseudo value for first replicate.}

\item{z.2}{pseudo value for second replicate.}

\item{mu}{the mean of the reproducible component.}

\item{sigma}{the standard deviation of the reproducible component.}

\item{rho}{the correlation coefficient of the reproducible component.}
}
\value{
the log likelihood of the binomial distribution with equal mean and signma on both dimensions.
}
\description{
density of binomial distribution with equal mean and sigma on both dimensions
}
