% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.step.2normal.discrete.R
\name{em.step.2normal.discrete}
\alias{em.step.2normal.discrete}
\title{The E-step and M-step of EM algorithm}
\usage{
em.step.2normal.discrete(z.1, z.2, m, mu, sigma, rho, p,
  count.as.singleton = 1, top.missing = F)
}
\arguments{
\item{z.1}{boundary of the unique and nonoverlapping categories based on pseduo values for the first replicate.}

\item{z.2}{boundary of the unique and nonoverlapping categories based on pseduo values for the second replicate.}

\item{m}{number of observations in each category.}

\item{mu}{a starting value for the mean of the reproducible component.}

\item{sigma}{a starting value for the standard deviation of the reproducible component.}

\item{rho}{a starting value for the correlation coefficient of the reproducible component.}

\item{p}{a starting value for the proportion of reproducible component.}

\item{count.as.singleton}{the count being seen as singleton. Default is 1.}

\item{top.missing}{if there are missing observations, top.missing=T.}
}
\value{
estimated parameters: p, rho, mu, sigma
}
\description{
The E-step and M-step of EM algorithm
}
