% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.IDR.discrete.R
\name{est.IDR.discrete}
\alias{est.IDR.discrete}
\title{EM algorithm to estimate the copula mixture model for discrete data}
\usage{
est.IDR.discrete(x, y, mu, sigma, rho, p, eps, n.missing, miss.sym,
  as.single.loglik, as.single.em, common.only = TRUE, labels = NULL)
}
\arguments{
\item{x}{replicate 1 : a vector of values from original observation}

\item{y}{replicate 2 : paired with \code{x}, a repeated measure having same length as \code{x}.}

\item{mu}{a starting value for the mean of the reproducible component.}

\item{sigma}{a starting value for the standard deviation of the reproducible component.}

\item{rho}{a starting value for the correlation coefficient of the reproducible component.}

\item{p}{a starting value for the proportion of reproducible component.}

\item{eps}{a small number to control convergence. For example, \code{esp=0.0001}.}

\item{n.missing}{\code{n.missing=0} indicates that we run the program as if there is no missing
values; \code{n.missing=}positive number: the program estimates missing values.}

\item{miss.sym}{symbol for missing value. For example, \code{miss.sym=0}.}

\item{as.single.loglik}{an integer n to control the precision of
 numberical integration. If n=1, the computation is exact.
If n>1, only integrate the bins with counts more than n,
 and treat bins with counts <= n as singletons in the
 likelihood computation}

\item{as.single.em}{similar to \code{as.single.loglik}, except that
it does the approximation in EM}

\item{common.only}{If \code{TRUE}, the computation will only use the
nonmissing entries. In this case, \code{n.missing}
should be set to 0. If \code{FALSE}, it will use
all the entries.}

\item{labels}{the cdf on the left side of the lower boundary of the bins}
}
\value{
\itemize{
  \item{para   }{estimated parameters: p, rho, mu, sigma, estimated n.missing.}
  \item{loglik  }{can be used for mapping category format back to original
                  e.g. \code{x.cat[level.factor[1]] == x[1], y.cat[level.factor[1]]==y[1]}.}
  \item{loglik.trace	 }{trajectory of log-likelihood.}
  \item{idr.cat  }{a numeric vector of the local idr for each category (i.e. estimated conditional probablility for each observation to belong to the irreproducible component).}
  \item{IDR.cat  }{a numerical vector of the expected irreproducible discovery rate for categories that are as irreproducible or more irreproducible than the given categories.}
  \item{idr.obs  }{a numeric vector of the local idr for each observation.}
  \item{IDR.obs  }{a numerical vector of the expected irreproducible discovery rate for observations that are as irreproducible or more irreproducible than the given observations.}
}
}
\description{
EM algorithm to estimate the copula mixture model for discrete data
}
\details{
EM to compute the latent structure
steps:
1. raw values are first transformed into pseudovalues
2. EM is used to compute the underlining structure, which is a mixture of two normals
}
\examples{

#load chip_seq data
data(chip_seq)
x = chip_seq[,1]
y = chip_seq[,2]

# Initiation
mu <- 2.6
sigma <- 1.3
rho <- 0.8
p <- 0.7
eps <- 0.001
n.missing <- 0

# Estimate parameters of mixture model
gidr.out <- est.IDR.discrete(x, y, mu, sigma, rho, p, eps, n.missing,
                            miss.sym = 0, as.single.loglik = 1,
                            as.single.em = 1, common.only=TRUE, labels=NULL)

names(gidr.out)
}
