% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.cat.xy.R
\name{get.cat.xy}
\alias{get.cat.xy}
\title{Convert original observations to emprical cdf}
\usage{
get.cat.xy(x.ori, y.ori, n.missing, miss.sym = 0, top.missing = TRUE,
  multi.fac = 1e-04)
}
\arguments{
\item{x.ori}{replicate 1 : a vector of values from original observation}

\item{y.ori}{replicate 2 : paired with \code{x.ori}, a repeated measure having same length with \code{x.ori}.}

\item{n.missing}{the number of observations that miss both \code{x} and \code{y}.}

\item{miss.sym}{the numerical symbol to represent a missing observation.
It needs to be smaller than the smallest observed value.}

\item{top.missing}{if missing value is present, \code{top.missing=T}; if not consider missing values, \code{top.missing=F}.}

\item{multi.fac}{the cdf on the left side of the lower boundary of the bins}
}
\value{
\itemize{
  \item{cat.xy:   }{a data.frame with \code{nrow} = number of unique combinations of \code{(x, y)}.
                   It has the following columns:
                   \item{x, y: }{original values, only unique combinations to define a category.}
                   \item{factor.x, factor.y: }{level of factors, only unique values.}
                   \item{x.cdf, y.cdf: }{empirical cdf for each category, this is the upper boundary.}
                   \item{x.cdf.lo, y.cdf.lo: }{empirical cdf for each category, lower boundary.
                    (i.e. previous upper boundary). To avoid instability at boundary, the lowest \code{x.cdf.lo} (\code{y.cdf.lo}) is
                   \code{1/(n.xy+1)} and the highest \code{x.cdf} (\code{y.cdf}) is \code{n.xy/(n.xy+1)}.}
                   \item{m: }{number of observations in each category.}
                   }
   \item{level.factor:  }{can be used for mapping category format back to original
                  e.g. \code{x.cat[level.factor[1]] == x[1], y.cat[level.factor[1]]==y[1]}.}
}
}
\description{
Convert original observations to emprical cdf
}
\details{
Convert \code{(x, y)} from original observations to emprical cdf and unique pairs.
}
\examples{
data(chip_seq)
x = chip_seq[,1]
y = chip_seq[,2]
get.cat.xy(x, y, 0, miss.sym=0, top.missing=TRUE, multi.fac=0.0001)
}
