% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.IDR.discrete.R
\name{get.IDR.discrete}
\alias{get.IDR.discrete}
\title{compute IDR for discrete categories}
\usage{
get.IDR.discrete(idr, cat.counts)
}
\arguments{
\item{idr}{local idr for each category.}

\item{cat.counts}{the number of observations in each category.}
}
\value{
a numerical vector of the expected irreproducible discovery rate for categories that are as irreproducible or more irreproducible than the given categories.
}
\description{
compute IDR for discrete categories
}
