% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pseudo.mix.map2.R
\name{get.pseudo.mix.map2}
\alias{get.pseudo.mix.map2}
\title{Convert cdf to pseudo values for one vector of cdfs of categories}
\usage{
get.pseudo.mix.map2(x.cdf, x.cdf.lo, mu, sigma, p)
}
\arguments{
\item{x.cdf}{the vector with upper boundary of each category.}

\item{x.cdf.lo}{the vector with lower boundary of each category.}

\item{mu}{the mean of the reproducible component.}

\item{sigma}{the standard deviation of the reproducible component.}

\item{p}{the proportion of reproducible component.}
}
\value{
the pseduo values for the vector of cdfs of categroies.
}
\description{
Convert cdf to pseudo values for one vector of cdfs of categories
}
\details{
Convert cdf to pseudo values for one vector of cdfs of categories.
The computation was done by first using unique values, then mapping back to
original vector to get a fast computation.
}
