% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pseudo.mix.R
\name{get.pseudo.mix}
\alias{get.pseudo.mix}
\title{Convert cdf to pseudo values for one cdf vector}
\usage{
get.pseudo.mix(x.cdf, mu, sigma, p)
}
\arguments{
\item{x.cdf}{the vector cdf for the observations.}

\item{mu}{the mean of the reproducible component.}

\item{sigma}{the standard deviation of the reproducible component.}

\item{p}{the proportion of reproducible component.}
}
\value{
the pseduo values for the vector of cdfs.
}
\description{
Convert cdf to pseudo values for one cdf vector
}
