% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pseudo.mix.x.R
\name{get.pseudo.mix.x}
\alias{get.pseudo.mix.x}
\title{Convert cdf to pseudo values for one cdf value}
\usage{
get.pseudo.mix.x(x.cdf, mu, sigma, p)
}
\arguments{
\item{x.cdf}{the cdf for a single observation.}

\item{mu}{the mean of the reproducible component.}

\item{sigma}{the standard deviation of the reproducible component.}

\item{p}{the proportion of reproducible component.}
}
\value{
the pseduo value for the cdf value.
}
\description{
Convert cdf to pseudo values for one cdf value
}
