% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.start.value.R
\name{get.start.value}
\alias{get.start.value}
\title{use a continuous version to find starting values of mu, sigma, rho and n.missing}
\usage{
get.start.value(x, y)
}
\arguments{
\item{x}{replicate one}

\item{y}{replicate two}
}
\value{
starting values generated
}
\description{
use a continuous version to find starting values of mu, sigma, rho and n.missing
}
