% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_factorial.R
\name{log_factorial}
\alias{log_factorial}
\title{compute the log factorial}
\usage{
log_factorial(n)
}
\arguments{
\item{n}{integer}
}
\value{
log factorial of the given integer
}
\description{
compute the log factorial
}
