% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_factorial_vec.R
\name{log_factorial_vec}
\alias{log_factorial_vec}
\title{compute a vector of log factorial}
\usage{
log_factorial_vec(n.vec)
}
\arguments{
\item{n.vec}{a vector of integers}
}
\value{
a vector of log factorial for given integers
}
\description{
compute a vector of log factorial
}
