% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.2binormal.discrete.R
\name{loglik.2binormal.discrete}
\alias{loglik.2binormal.discrete}
\title{compute log-likelihood for mixture of two discretized bivariate normals}
\usage{
loglik.2binormal.discrete(z.1.all, z.2.all, m.all, mu, sigma, rho, p,
  count.as.singleton = 1, top.missing = F)
}
\arguments{
\item{z.1.all}{pseudo values for first normal.}

\item{z.2.all}{pseduo values for second normal.}

\item{m.all}{number of observations in each category.}

\item{mu}{the mean of the reproducible component.}

\item{sigma}{the standard deviation of the reproducible component.}

\item{rho}{the correlation coefficient of the reproducible component.}

\item{p}{the proportion of reproducible component.}

\item{count.as.singleton}{the count being seen as singleton. Default is 1.}

\item{top.missing}{if there are missing observations, \code{top.missing=T}.}
}
\value{
log-likehood for mixture of two discretized bivariate normals.
}
\description{
compute log-likelihood for mixture of two discretized bivariate normals
}
