% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.gIDR.R
\name{select.gIDR}
\alias{select.gIDR}
\title{Select observations according to gIDR}
\usage{
select.gIDR(x, IDR.x, IDR.level)
}
\arguments{
\item{x}{a n by m numeric matrix, where m= num of replicates, n=num of observations. Numerical values representing the significance of the observations, where larger values represent higher significance, for example, -log(p-value). Currently, m=2.}

\item{IDR.x}{Irreproducibile discovery rate for each entry of x. It is computed from \code{est.IDR.discrete()}.}

\item{IDR.level}{IDR cutoff, a numerical value between [0, 1]. Typical value is 0.01 or 0.05.}
}
\value{
\itemize{
   \item{x:   }{selected reproducible observations given the cutoff.}
   \item{n:   }{number of selected reproducible observations.}#'
    }
}
\description{
Select observations that exceeding a given IDR level
}
